Move convert field boolean to DruidSqlValidator from BaseDruidSqlValidator
