+++ sql/src/main/java/org/apache/calcite/prepare/BaseDruidSqlValidator.java
@@ -31,8 +31,6 @@ import org.apache.calcite.sql.SqlOperatorTable;
  */
 public class BaseDruidSqlValidator extends CalciteSqlValidator
 {
-  private Boolean earliestLatestByConverted;
-
   public BaseDruidSqlValidator(
       SqlOperatorTable opTab,
       CalciteCatalogReader catalogReader,
@@ -41,16 +39,5 @@ public class BaseDruidSqlValidator extends CalciteSqlValidator
   )
   {
     super(opTab, catalogReader, typeFactory, validatorConfig);
-    earliestLatestByConverted = false;
-  }
-
-  public void setEarliestLatestByConverted()
-  {
-    this.earliestLatestByConverted = true;
-  }
-
-  public Boolean getEarliestLatestByConverted()
-  {
-    return earliestLatestByConverted;
   }
 }
