+++ sql/src/main/java/org/apache/druid/sql/calcite/planner/DruidSqlValidator.java
@@ -28,8 +28,11 @@ import org.apache.calcite.sql.SqlOperatorTable;
  * Druid extended SQL validator. (At present, it doesn't actually
  * have any extensions yet, but it will soon.)
  */
-class DruidSqlValidator extends BaseDruidSqlValidator
+public class DruidSqlValidator extends BaseDruidSqlValidator
 {
+
+  private Boolean earliestLatestByConverted;
+
   protected DruidSqlValidator(
       SqlOperatorTable opTab,
       CalciteCatalogReader catalogReader,
@@ -38,5 +41,16 @@ class DruidSqlValidator extends BaseDruidSqlValidator
   )
   {
     super(opTab, catalogReader, typeFactory, validatorConfig);
+    earliestLatestByConverted = false;
+  }
+
+  public Boolean getEarliestLatestByConverted()
+  {
+    return earliestLatestByConverted;
+  }
+
+  public void setEarliestLatestByConverted()
+  {
+    this.earliestLatestByConverted = true;
   }
 }
